refactor(mapping): split mappings into domain files; add helpers
